kick off notes interpreter
some refactoring:
* turn some easy-to-mess-up-via-copy-paste code into macros
* expose InputBuffer via IInstanceInterpreter (only Input reacts pleasantly to this)
* note value should be double
* expose Input::stream_ type in global namespace
* forgot to export InstanceInterpreter<Noise>

+++ jakmuse/parser_types.h
@@ -3,7 +3,7 @@
 
 struct Note {
     int duration;
-    float value;
+    double value;
 };
 
 struct PpValueList;
